move tour_points and encounters to api namespace
