+++ app/views/api/v0/encounters/create.json.jbuilder
