+++ app/views/api/v0/encounters/error.json.jbuilder
