+++ spec/controllers/api/v0/encounters_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe EncountersController, :type => :controller do
+RSpec.describe Api::V0::EncountersController, :type => :controller do
   render_views
 
   describe "POST create with tour" do
