+++ spec/controllers/api/v0/tour_points_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe TourPointsController, :type => :controller do
+RSpec.describe Api::V0::TourPointsController, :type => :controller do
   render_views
 
   describe "POST create" do
